+++ src/common/traits/IAttackable.cs
@@ -1,8 +1,22 @@
 namespace EternalJourney.Common.Traits;
 
+/// <summary>
+/// 攻撃可能インターフェース
+/// </summary>
 public interface IAttackable
 {
-    public double Atk { get; set; }
+    /// <summary>
+    /// 攻撃力
+    /// </summary>
+    public float Atk { get; set; }
+
+    /// <summary>
+    /// 攻撃可能フラグ
+    /// </summary>
     public bool CanAttack { get; set; }
+
+    /// <summary>
+    /// 攻撃
+    /// </summary>
     public void Attack();
 }
